Use -u user and base url instead of user@host
The software on the Avocent stuff is so old, that you cannot open a
https connection any more from a recent OS (like, Debian Buster).

These products are discontinued, there will be no newer software
updates. So, instead, we should just enable the http web server, and
disable https, and then connect over plain http, while not using
unsecured transport in the middle.

However, the program cannot do that currently, since it's hardcoding
https.  Morph the way in which user and host are specified a bit, so
that the user and a base url using either http or https can be
specified:

Before:
    ./pmCommand.py user@10.13.3.7

After:
    ./pmCommand.py -u admin http://10.13.3.7

If not specified, the default user is 'admin'.

Fixes: https://github.com/mendix/pmCommand.py/issues/4