Remove version number from code
Either the develop branch would have previous version set here, or it
has to stay UNRELEASED in the develop branch and then adjusted after
merging to master.

I don't really care about it, so let's remove it. Whoever will be using
this should be smart enough to be able to tell what exact code they're
running when reporting an issue.