Always set prompt to pmCommand#
Later, I want to be able to log in to two separate PDUs which are not
connected to an ACS, and disable or enable ports on any of them. In that
case it doesn't really make sense to put a list of ip addresses or
hostnames into the prompt. Just use the status command to see what
you're connected to.