Rewrite all the structures stuff
Ok, this is a pretty big change, but I can't really split it up and have
a working program on each intermediate commit without a lot of overhead.

Basically, I didn't really like how I did this in 2014, and it contained
a few things which would really be in the way of having this tool deal
with multiple PDUs to log into instead of just one PDU or ACS.

Changes:
- No custom mappings for field names, let's just use the ones from the
avocent xml stuff. This will make it a bit easier to debug when someone
else starts reverse engineering more things. This also makes the PDU and
Outlet class a lot simpler.
- Add a structures.DeviceFromXML class as super class for PDU and
Outlet, which contains the generic logic to just tetris field names,
column headers and actual values into the object. This replaces the old
load_array function in the module.
- No more class variables, so that we can later have multiple PMCommand
objects which use the PDU and Outlet structures. (maybe even combining
PDUs seen via an ACS with directly logging into another extra PDU)...
- Just use header and value instead of field, label, header, text, or
wherever all of the confusing names where it was stored before.
- Move print_table from cli to util.
- In Outlet, we previously already combined the pdu_id with
outlet_number to get the combined value in the outlet column (e.g.
'power2-213[3]'. This is the format that the outlet commands accept as
args (e.g. 'off power2-213[3],power4-213[3]'). Since we now just store
the attribute names and values as we get them from the xml, I added
output filter functionality to the print_table function, which can
provide a function to apply on the value of a field just before
displaying it.

Fixes: https://github.com/mendix/pmCommand.py/pull/3