Rework help function
inspect does not return modules in the order they are written down in the code.
Use a helper variable pref_order to keep the order we preferr.
Fetch the first line of the docstring from the method and use that.
This allows for additional docstring text without clobbering the help()
function.